requirements.txt
numpy>=1.24.1
pandas>=1.5.3
matplotlib>=3.6.3
networkx>=3.3
pyomo>=6.4.4
gurobi>=10.0.0